Fix test unit to work with standalone Hocon
