+++ src/HOCON.Tests/Extensions/IsStringTests.cs
@@ -24,7 +24,7 @@ namespace Hocon.Tests.Extensions
             }
         ";
 
-        public HoconRoot TestHocon => Parser.Parse(RawTestHocon);
+        public HoconRoot TestHocon => HoconParser.Parse(RawTestHocon);
 
         [Fact]
         public void IsString_should_detect_String_literals()
