+++ src/Hocon.Immutable.Tests/ApiTest.cs
@@ -56,7 +56,7 @@ namespace Hocon.Immutable.Tests
         [MemberData(nameof(ObjectArrayData))]
         public void CanIntelligentlyUseIndexerTypeToAccessMembers(string hocon)
         {
-            var config = Parser.Parse(hocon).ToHoconImmutable();
+            var config = HoconParser.Parse(hocon).ToHoconImmutable();
 
             Assert.Equal(1, config["a"][0]["a"].GetInt());
             Assert.Equal(2, config["a"][0]["b"].GetInt());
@@ -128,7 +128,7 @@ namespace Hocon.Immutable.Tests
 
   root_2 : 1234
 }";
-            var config = Parser.Parse(hocon).ToHoconImmutable();
+            var config = HoconParser.Parse(hocon).ToHoconImmutable();
 
 #pragma warning disable 618
             Assert.True(config.Value.GetBoolean("root.bool")); // legacy
