+++ src/Hocon.Tests/UnquotedString.cs
@@ -20,7 +20,7 @@ namespace Hocon.Tests
         public void ShouldPreserveWhitespacesInTheMiddle()
         {
             var hocon = $"a = literal{Whitespace.Whitespaces}value";
-            var config = Parser.Parse(hocon);
+            var config = HoconParser.Parse(hocon);
 
             Assert.Equal($"literal{Whitespace.Whitespaces}value", config.GetString("a"));
         }
@@ -36,7 +36,7 @@ namespace Hocon.Tests
         public void ShouldRemoveAllLeadingWhitespace()
         {
             var hocon = $"a = {Whitespace.Whitespaces}literal value";
-            var config = Parser.Parse(hocon);
+            var config = HoconParser.Parse(hocon);
 
             Assert.Equal("literal value", config.GetString("a"));
         }
@@ -45,7 +45,7 @@ namespace Hocon.Tests
         public void ShouldRemoveAllTrailingWhitespace()
         {
             var hocon = $"a = literal value{Whitespace.Whitespaces}";
-            var config = Parser.Parse(hocon);
+            var config = HoconParser.Parse(hocon);
 
             Assert.Equal("literal value", config.GetString("a"));
         }
