+++ src/examples/SimpleSubstitutions/Program.cs
@@ -19,7 +19,7 @@ root {
   simple-string = ${root.some-value} ""Hocon""  
 }
 ";
-            var config = Parser.Parse(hocon);
+            var config = HoconParser.Parse(hocon);
             var val = config.GetString("root.simple-string");
             Console.WriteLine("Hocon says: " + val);
             Console.ReadKey();
